local

+++ src/utils/axiosInstance.js
@@ -19,7 +19,7 @@ axiosInstance.interceptors.request.use((config) => {
 axiosInstance.interceptors.response.use(
   (response) => response,
   (error) => {
-    if (error.response?.status === 401) {
+    if (error.response?.status === 401 && !window.location.pathname.includes("login")) {
       sessionStorage.clear();
       window.location.href = "/login";
     }
